SPLINE: added a cast in dxf_spline_write().

+++ src/spline.c
@@ -874,7 +874,7 @@ dxf_spline_write
         }
         if (spline->paperspace == DXF_PAPERSPACE)
         {
-                fprintf (fp->fp, " 67\n%hd\n", DXF_PAPERSPACE);
+                fprintf (fp->fp, " 67\n%hd\n", (int16_t) DXF_PAPERSPACE);
         }
         fprintf (fp->fp, "  8\n%s\n", spline->layer);
         if (strcmp (spline->linetype, DXF_DEFAULT_LINETYPE) != 0)
